Refactor field setting
