+++ packages/cms/src/fields/slug/Setting/index.js
@@ -3,21 +3,22 @@ import React from "react";
 import Form from "fields/template/Setting/Form";
 import Name from "fields/template/Setting/Name";
 import Id from "fields/template/Setting/Id";
-import Default from "fields/template/Setting/Default";
 import Help from "fields/template/Setting/Help";
 import Required from "fields/template/Setting/Required";
 import Submit from "fields/template/Setting/Submit";
 
+import Reference from "./Reference";
+
 const Setting = ({ update, error, ...data }) => {
   const onSubmit = async (data) => {
     const field = {
-      type: "text",
+      type: "slug",
       id: data.id,
       name: data.name,
       validations: {
         required: data.required,
       },
-      default: data.default,
+      reference: data.reference,
       help: data.help,
     };
     update(field);
@@ -29,7 +30,7 @@ const Setting = ({ update, error, ...data }) => {
         id: data.id,
         name: data.name,
         required: data?.validations?.required,
-        default: data.default,
+        reference: data.reference,
         help: data.help,
       }}
       onSubmit={onSubmit}
@@ -38,7 +39,7 @@ const Setting = ({ update, error, ...data }) => {
       {error && <div className="text-danger font-weight-bold">{error}</div>}
       <Name />
       <Id data={data} />
-      <Default />
+      <Reference />
       <Help />
       <Required />
       <Submit data={data} />
